requirements.txt
requests
tqdm

# developing
python-dotenv
